requirements.txt
fastapi
uvicorn
requests
discord.py
google-cloud-secret-manager
openai